updates readme.md and adds examples
